Add requestBuilder parameter to CoilImage
