+++ imageloading-core/build.gradle
@@ -51,12 +51,6 @@ android {
         // We run a full lint analysis as build part in CI, so skip vital checks for assemble tasks
         checkReleaseBuilds false
     }
-
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-        }
-    }
 }
 
 afterEvaluate {
@@ -72,19 +66,6 @@ dependencies {
 
     implementation Libs.Kotlin.stdlib
     implementation Libs.Coroutines.android
-
-    androidTestImplementation Libs.junit
-    androidTestImplementation Libs.truth
-    androidTestImplementation Libs.mockk
-
-    androidTestImplementation Libs.mockWebServer
-
-    androidTestImplementation Libs.Coroutines.test
-
-    androidTestImplementation Libs.AndroidX.Compose.test
-    androidTestImplementation Libs.AndroidX.Compose.ui
-    androidTestImplementation Libs.AndroidX.Test.rules
-    androidTestImplementation Libs.AndroidX.Test.runner
 }
 
 apply plugin: "com.vanniktech.maven.publish"
